requirements.txt
asgiref==3.8.1
Django==4.2
gunicorn==22.0.0
packaging==24.1
python-dotenv==1.0.1
sqlparse==0.5.1
tzdata==2024.1
requests==2.32.3
django-autocomplete-light==3.11.0
